+++ client/src/pages/admin/AdminDashboard.jsx
@@ -7,8 +7,11 @@ import { getSubjects } from "../../services/subjects.ts";
 import { useState, useEffect } from "react";
 import { useUser } from "../../context/userContext";
 import { Icon } from "@iconify/react";
+import { Link } from "react-router-dom";
+import CountUp from 'react-countup';
 import Pagination from "../../components/ui/Pagination.jsx";
 import LoadingIcon from "../../components/ui/LoadingIcon.jsx";
+import { useGreeting } from "../../hooks/useGreeting.ts";
 
 function AdminDashboard() {
   const [masterlist, setMasterlist] = useState([]);
@@ -18,6 +21,7 @@ function AdminDashboard() {
   const { student } = useUser();
   const [resourcePage, setResourcePage] = useState(1);
   const [loading, setLoading] = useState(false);
+  const greetings = useGreeting();
 
   const fetchMasterlist = async () => {
     try {
@@ -83,26 +87,13 @@ function AdminDashboard() {
     name: subject.code,
     value: task.filter(t => t.subject.code === subject.code).length,
   })).filter(s => s.value > 0);
-
-  // PARTIAL
-  // const getGreeting = () => {
-  //   const time = new Date();
-  //   let t = time.getHours();
-  //   if (t >= 21 || t <= 4){
-  //     console.log("midnight");
-  //   }
-  // }
   
-  // getGreeting();
-
   return (
     <div className="bg-[#fafafa] min-h-screen flex">
       <Sidebar />
-
       <div className="flex-1 p-8">
-
-        <div className="mb-7 leading-12">
-          <p className="font-bold text-[2rem] font-[montserrat]">Hello love, <span className="font-[parisienne] font-bold pl-1 text-[2.2rem] text-[#387c39]">{student?.user?.name.split(" ").at(0)}!</span></p>
+        <div className="mb-7 leading-10">
+          <p className="font-bold text-[2rem] font-[montserrat]">{greetings}, <span className="font-[parisienne] font-bold pl-1 text-[2.2rem] text-[#387c39]">{student?.user?.name.split(" ").at(0)}!</span></p>
           <p className="text-gray-400 text-sm">Here's what's happening in your section.</p>
         </div>
 
@@ -113,7 +104,7 @@ function AdminDashboard() {
             </div>
             <div>
               <label className="text-gray-400 font-bold text-xs uppercase tracking-widest">Total Students</label>
-              <p className="text-[2rem] font-bold text-[#1B651B] leading-8">{masterlist.length}</p>
+              <p className="text-[2rem] font-bold text-[#1B651B] leading-8"><CountUp.default start={0} end={masterlist.length} duration={3}/></p>
             </div>
           </div>
 
@@ -123,7 +114,7 @@ function AdminDashboard() {
             </div>
             <div>
               <label className="text-gray-400 font-bold text-xs uppercase tracking-widest">Uploaded Tasks</label>
-              <p className="text-[2rem] font-bold text-[#185FA5] leading-8">{task.length}</p>
+              <p className="text-[2rem] font-bold text-[#185FA5] leading-8"><CountUp.default start={0} end={task.length} duration={3}/></p>
             </div>
           </div>
 
@@ -133,23 +124,24 @@ function AdminDashboard() {
             </div>
             <div>
               <label className="text-gray-400 font-bold text-xs uppercase tracking-widest">Uploaded Resources</label>
-              <p className="text-[2rem] font-bold text-[#BA7517] leading-8">{resources.length}</p>
+              <p className="text-[2rem] font-bold text-[#BA7517] leading-8"><CountUp.default start={0} end={resources.length} duration={3}/></p>
             </div>
           </div>
         </div>
 
         <div className="bg-white rounded-xl border border-gray-200 p-5 mt-5 flex flex-col">
-          <p className="font-bold text-[1rem] mb-4">Recent Resources</p>
+          <p className="font-bold text-[1rem] mb-4">Recent Uploaded Resources</p>
 
           {/* Recent Resources */}
-          <div className="h-[285px] flex flex-col justify-start">
+          <div className="h-71.25 flex flex-col justify-start">
             {loading ? (
               <div className="flex justify-center items-center flex-1">
                 <LoadingIcon dimensions="w-10 h-10" />
               </div>
             ) : (
               paginatedResources.length > 0 ? (
-                paginatedResources.map((r) => (
+                paginatedResources.map((r, i) => (
+                  <Link key={i} to={"/admin/resources"} >
                   <div key={r.id} className="flex items-center gap-3 px-3 py-2 rounded-md border-b border-gray-100 transition-all duration-100 hover:bg-gray-100">
                     <div className="bg-[#faeeda] rounded-lg p-2">
                       <Icon icon="mdi:file-document-outline" width="20" className="text-[#BA7517]" />
@@ -160,15 +152,15 @@ function AdminDashboard() {
                     </div>
                     <p className="text-gray-400 text-xs">{new Date(r.createdAt).toLocaleDateString()}</p>
                   </div>
+                  </Link>
                 ))
-            ) : (
-              <div className="flex justify-center items-center h-full">
-                <p className="text-gray-400">No resources uploaded yet.</p>
-              </div>
-            ))
-          }
+              ) : (
+                <div className="flex justify-center items-center h-full">
+                  <p className="text-gray-400">No resources uploaded yet.</p>
+                </div>
+              )
+            )}
           </div>
-
           <Pagination
             currentPage={resourcePage}
             totalPages={totalResourcePages}