+++ client/src/pages/sections/ClassResources.jsx
@@ -8,13 +8,14 @@ import Button from "../../components/ui/Button";
 import Modal from "../../components/ui/Modal";
 import { useUser } from "../../context/userContext";
 import UserIcon from "../../components/common/UserIcon";
-
+import LoadingIcon from "../../components/ui/LoadingIcon.jsx";
 
 export default function ClassResources() {
   const [resources, setResources] = useState([]);
   const [subjects, setSubjects] = useState([]);
   const { student } = useUser();
   const [modalOpen, setModalOpen] = useState(false);
+  const [loading, setLoading] = useState(false);
 
   // Add 
   const [title, setTitle] = useState("");
@@ -28,13 +29,17 @@ export default function ClassResources() {
   });
 
   const fetchResources = async () => {
+    setLoading(true);
     try {
       const data = await getResources();
       setResources(data);
     } catch (error) {
       console.log(error);
-    } 
+    } finally {
+      setLoading(false);
+    }
   }
+  
   const fetchSubjects = async () => {
     try {
       const data = await getSubjects();
@@ -96,27 +101,33 @@ export default function ClassResources() {
         <p className="font-bold text-[1.7rem] font-[montserrat] leading-7">Class Resources</p>
         <p className="text-[1rem]">Collaborate and exchange notes with your fellow Kabsuhenyos.</p>
       </div>
-      <div className="bg-white w-full mt-5 border border-gray-200 rounded-xl h-140 p-4 flex flex-col">
-        <div className="h-140 overflow-y-auto flex flex-col">
-          {resources.length > 0 ? (
-            <div className="grid grid-cols-5 gap-4 items-center">
-              {resources.map((resource) => (
-                <ResourceCard
-                  key={resource.id}
-                  title={resource.title}
-                  subject={resource.subject.code}
-                  uploadedBy={resource.uploadedBy.name}
-                  fileUrl={resource.fileUrl}
-                />
-              ))}
+      <div className="bg-white w-full mt-5 border border-gray-200 rounded-xl h-150 p-4 flex flex-col">
+        <div className="h-150 overflow-y-auto flex flex-col">
+          {loading ? (
+            <div className="flex justify-center items-center flex-1">
+              <LoadingIcon dimensions="w-10 h-10" />
             </div>
           ) : (
+            resources.length > 0 ? (
+              <div className="grid grid-cols-5 gap-4 items-center">
+                {resources.map((resource) => (
+                  <ResourceCard
+                    key={resource.id}
+                    title={resource.title}
+                    subject={resource.subject.code}
+                    uploadedBy={resource.uploadedBy.name}
+                    fileUrl={resource.fileUrl}
+                  />
+                ))}
+              </div>
+            ) : (
             <div className="flex justify-center items-center flex-1">
               <p className="text-gray-400">There's no resources uploaded yet.</p>
             </div>
+            )
           )}
         </div>
-        <div className="flex justify-center items-center p-5">
+        <div className="flex justify-center items-center pt-3">
           <Button
             text="+ Upload Resources"
             onClick={() => setModalOpen(true)}
@@ -169,7 +180,7 @@ export default function ClassResources() {
               <p className="text-red-500 text-xs">{error.subject}</p>
             )}
 
-            {/* File to be uploaded | only accepts .PDF*/}
+            {/* File to be uploaded */}
             <label className="text-xs font-bold mb-1 mt-2">File <span className="text-red-400">*</span></label>
             <input 
               onChange={(e) => {